Inverse of nullable not storing empty array([])
In the documentation, it states that `If the field is nullable an empty select will result in null else an empty array ([]) is stored`  but this is nor working as expected. Where nullable is not set it is still trying to store null which results in NULL constraint error. I am submitting a fix for this.